requirements.txt
numpy
opencv-python
tensorflow
Pillow
imutils
tkinter
